Partial commit: Battery code, Cont-A2C, and cleanup
Added the first part of a Battery model to be included in our scenario, also a better approach to do the A2C (with causality trick) algorithm coded by scratch. Also removed unuseful code lines mainly in the microgrid.py file, the dataframes to plot are not useful, the logging feature isn't useful neither. Fixed the requirements.txt file.